+++ appveyor.yml
@@ -23,9 +23,6 @@ init:
 build: off
 
 install:
-  # Upgrade to the latest version of pip to avoid it displaying warnings
-  # about it being out of date.
-  - pip install -U pip
   # Install tox to run tests
   - pip install tox
