+++ deezer/__init__.py
@@ -4,14 +4,13 @@ Implements several classes to interact with all types of Deezer objects,
 do some searches, and build application written in Python on top of it
 
 """
-from deezer.asynchronous import AsyncClient
 from deezer.client import Client
 from deezer.resources import Album, Resource, Artist, Playlist
 from deezer.resources import Genre, Track, User, Comment, Radio
 
 
 __version__ = '0.7.0'
-__all__ = ['AsyncClient', 'Client', 'Resource', 'Album',
+__all__ = ['Client', 'Resource', 'Album',
            'Artist', 'Genre', 'Playlist', 'Track', 'User',
            'Comment', 'Radio']
