+++ docs/source/api_reference/client.rst
