+++ docs/source/api_reference/contrib/tornado.rst
@@ -1,0 +1,5 @@
+Async tornado client
+--------------------
+
+.. automodule:: deezer.contrib.tornado
+    :members:
