+++ docs/source/api_reference/toc.rst
@@ -3,6 +3,6 @@ API Reference
 
 .. toctree::
 
-    asynchronous
     client
     resources
+    contrib/tornado
