--- docs/source/asynchronous.rst
@@ -1,5 +1,0 @@
-Async module
-------------
-
-.. automodule:: deezer.asynchronous
-    :members:
